+++ R/pattern.R
@@ -140,10 +140,11 @@ group_pattern = function(pattern) {
 # automatically detect the chunk patterns
 detect_pattern = function(text, ext) {
   if (!missing(ext)) {
+    ext = tolower(ext)
     if (ext %in% c('rnw', 'snw', 'stex')) return('rnw')
     if (ext == 'brew') return('brew')
     if (ext %in% c('htm', 'html', 'rhtm', 'rhtml')) return('html')
-    if (ext %in% c('rmd', 'rmarkdown', 'markdown', 'md')) return('md')
+    if (ext %in% c('rmd', 'rmarkdown', 'markdown', 'md', 'qmd')) return('md')
     if (ext %in% c('rst', 'rrst')) return('rst')
     if (ext %in% c('asciidoc', 'rasciidoc', 'adoc', 'radoc')) return('asciidoc')
   }